requirements.txt
grpcio-tools
protobuf == 5.27.2